Initial support for global operator hints
This adds a rough implementation of the global configurations known from Postgres, i.e. stuff like enable_nestloop = 'on'. We integrate these settings as part of the hint block, which has the advantage of only affecting the current query. The syntax is
Set(enable_hashjoin = 'off')
The operators currently use the same names as the well-known GUC settings from Postgres.

Note that this is a "move fast and break things" commit because we needed to comment-out an assertion from the normal DuckDB implementation. We don't understand the DuckDB internals deep enough yet to correctly judge whether this is actually safe, but it seemed to work in our testing.